requirements.txt
aiosmtpd
celery
requests
gunicorn
python-dotenv
pan-python
redis